chore: removes unnecessary modules nesting around "validateMessage"
