+++ lib/index.js
@@ -1,4 +1,4 @@
-const validate = require('./next/validate');
+const { validate } = require('./next/validate');
 
 module.exports = {
   validate
