+++ lib/next/test/integration/validate.test.js
@@ -1,5 +1,5 @@
 const { assert } = require('chai');
-const { validateMessage } = require('../../validateMessage');
+const { validate } = require('../../validate');
 
 const validator = (obj, expected) => {
   it(`has "${expected}" validator`, () => {
@@ -21,7 +21,7 @@ const noErrors = (obj) => {
   });
 };
 
-describe('validateMessage', () => {
+describe('validate', () => {
   describe('with matching requests', () => {
     const request = {
       method: 'POST',
@@ -30,7 +30,7 @@ describe('validateMessage', () => {
       },
       body: '{ "foo": "bar" }'
     };
-    const result = validateMessage(request, request);
+    const result = validate(request, request);
 
     it('returns validation result object', () => {
       assert.isObject(result);
@@ -73,7 +73,7 @@ describe('validateMessage', () => {
   });
 
   describe('with non-matching requests', () => {
-    const result = validateMessage(
+    const result = validate(
       {
         method: 'PUT',
         headers: '',
@@ -162,7 +162,7 @@ describe('validateMessage', () => {
       },
       body: '{ "foo": "bar" }'
     };
-    const result = validateMessage(response, response);
+    const result = validate(response, response);
 
     it('returns validation result object', () => {
       assert.isObject(result);
@@ -217,7 +217,7 @@ describe('validateMessage', () => {
         'Accept-Language': 'en-US'
       }
     };
-    const result = validateMessage(expectedResponse, realResponse);
+    const result = validate(expectedResponse, realResponse);
 
     it('returns validation result object', () => {
       assert.isObject(result);
@@ -279,7 +279,7 @@ describe('validateMessage', () => {
   });
 
   describe('with non-matching headers', () => {
-    const result = validateMessage(
+    const result = validate(
       {
         statusCode: 404,
         headers: {
@@ -346,7 +346,7 @@ describe('validateMessage', () => {
   });
 
   describe('always validates expected properties', () => {
-    const result = validateMessage(
+    const result = validate(
       {
         statusCode: 200,
         headers: {
